refactor: Split the models.py and creates a models directory
