+++ learning_environment/models/__init__.py
@@ -1,0 +1,3 @@
+from .users import User, Profile
+from .lessons import Lesson, Task
+from .statuses import ProfileSeriesLevel, LearnerStatus, UserLesson, Solution
